requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0
Pillow>=9.4.0
tqdm>=4.65.0
matplotlib>=3.7.0
Flask>=2.2.2
flask-cors
albumentations>=1.3.0
facenet-pytorch>=2.5.2

